Setup Docker and CI/CD pipelines for Rust project
The Dockerfile, .dockerignore file, a Gitlab CI/CD pipeline, and Github action are added in this commit to set up automation for building and testing the Rust application. Additionally, a corresponding CI/CD script and a basic README have been created as a part of the setup process. An update to the tax_calculator project settings is also included.